feat: moved states to package
